+++ CHANGELOG.md
@@ -16,7 +16,7 @@ alt="codacy coverage" height="20">
 <img src="https://img.shields.io/github/issues/LiphuCL/liphu-ui.svg?&style=flat&colorB=fabe46" alt="issues open"
 height="20">
 </a>
-<img src="https://img.shields.io/badge/commits%20to%20be%20deployed-1-3b79db.svg"
+<img src="https://img.shields.io/badge/commits%20to%20be%20deployed-6-3b79db.svg"
 alt="commits to be deployed">
 <a href="https://www.npmjs.com/package/@liphu/liphu-ui">
 <img src="https://img.shields.io/npm/v/@liphu/liphu-ui.svg?colorB=3b79db" alt="npm version" height="20">
@@ -33,7 +33,14 @@ All notable changes to this project will be documented in this file. :grin:
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 0.1.0 - 26 June 2020
+## 0.1.0 - 18 February 2021
+
+### Added
+
+- :star2: Added: icon component [`26dabec`](https://github.com/LiphuCL/liphu-ui/commit/26dabec80c0ff23eeebe44d8a5bf948488a562ab)
+- :star2: Added: card component [`191a65e`](https://github.com/LiphuCL/liphu-ui/commit/191a65e682e9ab4a26d54fb126f191064af8b845)
+- :star2: Added: avatar component [`918fc81`](https://github.com/LiphuCL/liphu-ui/commit/918fc8178783130a4134e5b232283bd716dc4385)
+- :star2: Added: divider component [`438de1b`](https://github.com/LiphuCL/liphu-ui/commit/438de1b614d87d0050f4d00f64f05b3fb462c503)
 
 ---
 <sub><sup>*This is an auto generated changelog. If you have any problems or suggestions, create an issue.* :blush: